requirements.txt
pydantic>=2.0.0
numpy>=1.24.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.0.0
gradio==5.25.1
plotly>=5.0.0